Refactoring diffImpl
